Removed unneeded folder files
